Refactor: Move UI-related files to core.ui package.
Moved files related to UI components, navigation, screens, themes, and view models from the `com.currency.rateman.ui` package to `com.currency.rateman.core.ui`. This refactoring improves the project structure by centralizing core UI elements.
