+++ app/src/main/java/com/currency/rateman/core/ui/components/BottomNavBar.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.components
+package com.currency.rateman.core.ui.components
 
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -12,7 +12,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import com.currency.rateman.ui.navigation.BottomNavItem
+import com.currency.rateman.core.ui.navigation.BottomNavItem
 
 @Composable
 fun BottomNavBar(
