+++ app/src/main/java/com/currency/rateman/core/ui/components/ThemeSettingItem.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.components
+package com.currency.rateman.core.ui.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -8,20 +8,19 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
 import androidx.compose.ui.Alignment
 import androidx.annotation.DrawableRes
 import androidx.compose.ui.res.painterResource
-import androidx.navigation.NavHostController
-import com.currency.rateman.ui.navigation.Routes
+import androidx.compose.ui.res.stringResource
+import com.currency.rateman.R
 
 @Composable
-fun <T : Enum<T>> LanguageSettingItem(
+fun <T : Enum<T>> ThemeSettingItem(
     label: String,
     value: String,
+    options: List<T>,
+    onValueChange: (T) -> Unit,
     @DrawableRes iconRes: Int,
-    navController: NavHostController,
     modifier: Modifier = Modifier
 ) {
     Box(
@@ -31,7 +30,6 @@ fun <T : Enum<T>> LanguageSettingItem(
                 MaterialTheme.colorScheme.surface,
                 shape = RoundedCornerShape(12.dp)
             )
-            .clickable { navController.navigate(Routes.Language.route) }
             .padding(12.dp)
     ) {
         Row(
@@ -58,20 +56,34 @@ fun <T : Enum<T>> LanguageSettingItem(
                 )
             }
             Row(
-              verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
             ) {
-                Text(
-                    text = value,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurface,
-                    modifier = Modifier.padding(top = 4.dp)
-                )
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.KeyboardArrowRight,
-                    contentDescription = "Expand",
-                    tint = MaterialTheme.colorScheme.onSurface,
-                    modifier = Modifier.size(24.dp)
-                )
+                options.forEach { option ->
+                    val isSelected = option.name == value
+
+                    val displayText = when (option.name) {
+                        "DARK" -> stringResource(R.string.theme_dark)
+                        "LIGHT" -> stringResource(R.string.theme_light)
+                        else -> option.name.lowercase().replaceFirstChar { it.uppercase() }
+                    }
+                    Box(
+                        modifier = Modifier
+                            .background(
+                                color = if (isSelected) MaterialTheme.colorScheme.background else MaterialTheme.colorScheme.tertiary,
+                                shape = RoundedCornerShape(8.dp)
+                            )
+                            .clickable {
+                                onValueChange(option)
+                            }
+                            .padding(horizontal = 12.dp, vertical = 6.dp)
+                    ) {
+                        Text(
+                            text = displayText,
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurface
+                        )
+                    }
+                }
             }
         }
     }
