+++ app/src/main/java/com/currency/rateman/core/ui/navigation/AppRouter.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.navigation
+package com.currency.rateman.core.ui.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -10,13 +10,13 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.currency.rateman.R
-import com.currency.rateman.ui.screens.CurrencyScreen
-import com.currency.rateman.ui.screens.LanguageScreen
+import com.currency.rateman.core.ui.screens.BaseCurrencyScreen
+import com.currency.rateman.core.ui.screens.CurrencyScreen
+import com.currency.rateman.core.ui.screens.LanguageScreen
+import com.currency.rateman.core.ui.screens.SettingsScreen
+import com.currency.rateman.core.ui.screens.Splashscreen
 import com.currency.rateman.provider.ui.screen.ProviderDetailsScreen
-import com.currency.rateman.ui.screens.SettingsScreen
 import com.currency.rateman.provider.ui.screen.RatesScreen
-import com.currency.rateman.ui.screens.Splashscreen
-import com.currency.rateman.ui.screens.BaseCurrencyScreen
 
 @Composable
 fun AppRouter() {
