+++ app/src/main/java/com/currency/rateman/core/ui/screens/LanguageScreen.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.screens
+package com.currency.rateman.core.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -29,7 +29,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavHostController
 import com.currency.rateman.di.navigation.sharedKoinNavViewModel
-import com.currency.rateman.ui.viewmodels.SettingsViewModel
+import com.currency.rateman.core.ui.viewmodels.SettingsViewModel
 import com.currency.rateman.R
 import com.currency.rateman.core.data.model.enums.LanguageCode
 import java.util.Locale
