+++ app/src/main/java/com/currency/rateman/core/ui/screens/Splashscreen.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.screens
+package com.currency.rateman.core.ui.screens
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
