+++ app/src/main/java/com/currency/rateman/core/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.theme
+package com.currency.rateman.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
