+++ app/src/main/java/com/currency/rateman/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.ui.theme
+package com.currency.rateman.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
