+++ app/src/main/java/com/currency/rateman/di/KoinModules.kt
@@ -8,10 +8,10 @@ import com.currency.rateman.core.data.repository.SettingsRepository
 import com.currency.rateman.core.data.repository.SettingsRepositoryImpl
 import com.currency.rateman.provider.domain.usecase.GetAllProvidersUseCase
 import com.currency.rateman.provider.domain.usecase.GetAllProvidersUseCaseImpl
-import com.currency.rateman.ui.viewmodels.CurrencyViewModel
+import com.currency.rateman.core.ui.viewmodels.CurrencyViewModel
 import com.currency.rateman.provider.ui.viewmodel.ProviderDetailViewModel
 import com.currency.rateman.provider.ui.viewmodel.ProviderListViewModel
-import com.currency.rateman.ui.viewmodels.SettingsViewModel
+import com.currency.rateman.core.ui.viewmodels.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
