+++ app/src/main/java/com/currency/rateman/provider/ui/screen/ProviderListScreen.kt
@@ -12,16 +12,16 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.currency.rateman.ui.navigation.BottomNavItem
+import com.currency.rateman.core.ui.navigation.BottomNavItem
 import com.currency.rateman.provider.ui.viewmodel.ProviderListViewModel
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import com.currency.rateman.di.navigation.sharedKoinNavViewModel
-import com.currency.rateman.ui.components.BottomNavBar
+import com.currency.rateman.core.ui.components.BottomNavBar
 import com.currency.rateman.provider.ui.components.ProvidersList
-import com.currency.rateman.ui.components.SearchAndFilterHeader
+import com.currency.rateman.core.ui.components.SearchAndFilterHeader
 
 @Composable
 fun RatesScreen(
